+++ include/kanji_tools/kanji/NonLinkedKanji.h
@@ -25,7 +25,7 @@ protected:
   // constructor used by 'CustomFileKanji': has 'meaning' and 'reading'
   NonLinkedKanji(const Data& d, const std::string& name, const Radical& radical,
                  const std::string& meaning, const std::string& reading,
-                 int strokes, const Ucd* u)
+                 size_t strokes, const Ucd* u)
       : Kanji(name, d.getCompatibilityName(name), radical, strokes,
               d.getMorohashiId(u), d.getNelsonIds(u), d.getPinyin(u)),
         _meaning(meaning), _reading(reading) {}
@@ -33,7 +33,7 @@ protected:
   // constructor used by 'CustomFileKanji' and 'UcdFileKanji': has 'reading' and
   // looks up 'meaning'
   NonLinkedKanji(const Data& d, const std::string& name, const Radical& rad,
-                 const std::string& reading, int strokes, const Ucd* u)
+                 const std::string& reading, size_t strokes, const Ucd* u)
       : NonLinkedKanji(d, name, rad, d.ucd().getMeaning(u), reading, strokes,
                        u) {}
 private:
