+++ src/kanji/kanji.cpp
@@ -6,8 +6,9 @@
 namespace kanji_tools {
 
 Kanji::Kanji(const std::string& name, const OptString& compatibilityName,
-             const Radical& radical, int strokes, const OptString& morohashiId,
-             const NelsonIds& nelsonIds, const OptString& pinyin)
+             const Radical& radical, size_t strokes,
+             const OptString& morohashiId, const NelsonIds& nelsonIds,
+             const OptString& pinyin)
     : _name(name),
       _nonVariantName(MBChar::optionalWithoutVariationSelector(name)),
       _compatibilityName(compatibilityName), _radical(radical),
